# Conflicts:
#	.gradle/8.10/checksums/checksums.lock
#	.gradle/8.10/executionHistory/executionHistory.bin
#	.gradle/8.10/executionHistory/executionHistory.lock
#	.gradle/8.10/expanded/expanded.lock
#	.gradle/8.10/fileHashes/fileHashes.bin
#	.gradle/8.10/fileHashes/fileHashes.lock
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.gradle/buildOutputCleanup/cache.properties
#	.gradle/buildOutputCleanup/outputFiles.bin
#	.gradle/nb-cache/subprojects.ser
#	build.gradle
#	settings.gradle

+++ src/main/java/modelo/Asignatura.java
@@ -50,11 +50,11 @@ public class Asignatura extends AbstractEntity {
         return nombre;
     }
 
-    public short getCreditos() {
+    public byte getCreditos() {
         return creditos;
     }
 
-    public short getIntensidad() {
+    public byte getIntensidad() {
         return intensidad;
     }
